+++ utils/RCBot2_meta/bot_mtrand.h
@@ -90,7 +90,7 @@ public:
   void operator=(const MTRand_int32&) = delete; // assignment operator not defined
 // the two seed functions
 static void seed(unsigned long); // seed with 32 bit integer
-  void seed(const unsigned long*, size_t size) const; // seed with array
+  void seed(const unsigned long*, int size) const; // seed with array
 // overload operator() to make this a generator (functor)
   unsigned long operator()() { return rand_int32(); }
 // 2007-02-11: made the destructor virtual; thanks "double more" for pointing this out
